feat: data Network 모듈 추가
