+++ local.properties
@@ -8,3 +8,4 @@
 KAKAO_API_KEY="17256ef458637114ac2cd1d6e2cd44c2"
 KAKAO_API_KEY_TEST="5c6add2f5c574f5575beb34228d6a789"
 sdk.dir=/Users/choijunghyeon/Library/Android/sdk
+BASE_URL = "http://beginvegan.p-e.kr/"
